refactoring day13
