+++ 2019/src/lib.rs
@@ -12,3 +12,4 @@ pub mod day09;
 pub mod day10;
 pub mod day11;
 pub mod day12;
+pub mod day13;
