+++ 2019/tests/day13_test.rs
@@ -1,0 +1,14 @@
+#[cfg(test)]
+mod tests {
+    use aoc2019::day13;
+
+    #[test]
+    fn part1_works() {
+        assert_eq!(193, day13::part1());
+    }
+
+    #[test]
+    fn part2_works() {
+        assert_eq!(10547, day13::part2());
+    }
+}
